backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
playwright==1.41.0
beautifulsoup4==4.12.3
anthropic==0.21.0
groq==0.4.2
python-dotenv==1.0.1
pydantic==2.6.1
httpx==0.26.0
python-multipart==0.0.9
pandas==2.2.0
openpyxl==3.1.2
aiofiles==23.2.1
openai>=1.12.0
